Changed editIssueAttribute
